PtAudioBin: fix volume
The volume changed sometimes when switching to ASR and back to playback
again.

+++ libparlatype/src/gst/gstptaudiobin.h
@@ -48,7 +48,7 @@ struct _GstPtAudioBin
 	/* properties */
 	gboolean    player;
 	gboolean    asr;
-	gfloat      volume;
+	gdouble     volume;
 	gboolean    mute;
 };
